refactor(media): split the load-progress and torch-ops subsystems out of embedder.py
`vtscore/media/embedder.py` was 1513 lines, twice the next-largest module in
`vtscore/media/`, and only its last ~620 were the `MediaEmbedder` ABC that 20-odd
embedder modules subclass. Everything before it was two self-contained
subsystems the ABC itself never touches.

- `vtscore/media/torch_ops.py` — `extract_tensor`, `to_compute_device`,
  `embed_autocast`, `to_model_inputs`, `to_float32`.
- `vtscore/media/load_progress.py` — `timed_progress`, `IMPORT_MODULE_ESTIMATES`,
  `embedder_load_setup`, `hf_token`, `load_pretrained_local_first`, and the
  reference-counted monkey-patch machinery behind `intercept_tqdm_progress` /
  `intercept_weight_loading_progress`.

Pure code movement: no behaviour, no signatures, no call sites changed.
`vtscore.media.embedder` re-exports every public name, so the third-party
embedder import documented in `vtscore/docs/extending/embedders.md` keeps
working unchanged.

The per-embedder `_ProgressSlot` / `_ThreadLocalProgress` descriptor stays with
the ABC rather than moving to `load_progress`: it is the implementation of
`MediaEmbedder._on_progress`, not part of the load-time interception.

Two test patch targets follow the code (`vtscore.media.load_progress.time.sleep`,
`...._hub_metadata_preflight`), and the three docs that credit a file path for
these helpers now name the file that defines them.
